+++ rollup.config.js
@@ -5,7 +5,7 @@
 //  `node_modules` path for the sake of the browser, but keeping
 //  in event we can use import paths later
 import nodeResolve from '@rollup/plugin-node-resolve';
-import commonjs from 'rollup-plugin-commonjs';
+import commonjs from '@rollup/plugin-commonjs';
 
 /**
  * @external RollupConfig